KAFKA-14760: Move ThroughputThrottler from tools to clients
