+++ clients/src/main/java/org/apache/kafka/common/utils/ThroughputThrottler.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.tools;
+package org.apache.kafka.common.utils;
 
 
 /**
